Move getPhoto code to Utilities
Shared utilities and getPhoto in cameraWidget

Update translations

Review fixes
