+++ trikControl/src/utilities.cpp
@@ -16,7 +16,9 @@
 
 #include "QsLog.h"
 
-QImage Utilities::imageFromBytes(const QVector<int32_t> &array, int width, int height, const QString &format){
+using namespace trikControl;
+
+QImage  Utilities::imageFromBytes(const QVector<int32_t> &array, int width, int height, const QString &format){
 	// QImage requires 32-bit aligned scan lines
 	// Helper function to convert data
 	uchar *formattedData = nullptr;
@@ -27,7 +29,7 @@ QImage Utilities::imageFromBytes(const QVector<int32_t> &array, int width, int h
 		}
 		auto scanLineSize = static_cast<int>((static_cast<unsigned>(perLine + 3)) & 0xFFFFFFFC);
 		auto dst = formattedData = new uchar[scanLineSize * height];
-		for (auto src = array.begin(); src < array.end(); src += perLine) {
+		for (auto src = array.begin(); src < array.end(); src += width) {
 			dst = std::copy(src, src + perLine, dst);
 			dst += scanLineSize - perLine;
 		}
@@ -37,7 +39,8 @@ QImage Utilities::imageFromBytes(const QVector<int32_t> &array, int width, int h
 
 	if (!format.compare("rgb32", Qt::CaseInsensitive)) {
 		fmt = QImage::Format_RGB32;
-		copyAligned(4 * width);
+		formattedData = new uchar[width * height]; // RGB32 alligned by default just needs to copy
+		formattedData = std::copy(array.begin(), array.end(), formattedData);
 	} else if (!format.compare("rgb888", Qt::CaseInsensitive)) {
 		fmt = QImage::Format_RGB888;
 		copyAligned(3 * width);
@@ -54,4 +57,51 @@ QImage Utilities::imageFromBytes(const QVector<int32_t> &array, int width, int h
 		static auto cleanUp = [](void *p) { if (p) delete [](static_cast<decltype (formattedData)>(p)); };
 		return QImage(formattedData, width, height, fmt, cleanUp, formattedData);
 	}
+}
+
+static inline int32_t getMedian(uint8_t &a, uint8_t &b, uint8_t &c, uint8_t &d)
+{
+	if (a > b)
+		std::swap(a, b);
+	if (c > d)
+		std::swap(c, d);
+	if (a > c)
+		std::swap(a, c);
+	if (b > d)
+		std::swap(b, d);
+	return (static_cast<int32_t>(b) + c) >> 1;
+}
+
+QVector<int32_t> Utilities::rescalePhoto(const QVector<uchar> data)
+{
+	QVector<int32_t> result;
+	result.reserve(data.size() / 3); //Repack RGB88 from 3 x uint8_t into int32_t
+	constexpr auto IMAGE_WIDTH = 320;
+	constexpr auto IMAGE_HEIGHT = 240;
+	if (data.size() >= IMAGE_WIDTH * IMAGE_HEIGHT * 3) {
+		for(int row = 0; row < IMAGE_HEIGHT; row += 2) {
+			for(int col = 0; col < IMAGE_WIDTH; col += 2) {
+				auto row1 = &data[(row * IMAGE_WIDTH + col) * 3];
+				auto row2 = row1 + IMAGE_WIDTH*3;
+				auto r1 = row1[0];
+				auto g1 = row1[1];
+				auto b1 = row1[2];
+				auto r2 = row1[3];
+				auto g2 = row1[4];
+				auto b2 = row1[5];
+				auto r3 = row2[0];
+				auto g3 = row2[1];
+				auto b3 = row2[2];
+				auto r4 = row2[3];
+				auto g4 = row2[4];
+				auto b4 = row2[5];
+
+				result.push_back((getMedian(r1, r2, r3, r4) << 16)
+					| (getMedian(g1, g2, g3, g4) << 8)
+					| getMedian(b1, b2, b3, b4));
+			}
+		}
+	}
+
+	return result;
 }
