+++ trikScriptRunner/src/scriptExecutionControl.cpp
@@ -23,6 +23,8 @@
 #include <QRandomGenerator>
 #include <QsLog.h>
 
+#include "utilities.h"
+
 using namespace trikScriptRunner;
 
 ScriptExecutionControl::ScriptExecutionControl(trikControl::BrickInterface &brick): mBrick(brick) {
@@ -39,7 +41,7 @@ void ScriptExecutionControl::reset()
 	mInEventDrivenMode = false;
 	emit stopWaiting();
 	for (auto &&timer : mTimers) {
-		QMetaObject::invokeMethod(timer, "stop", Qt::QueuedConnection);
+		QMetaObject::invokeMethod(timer, &QTimer::stop, Qt::QueuedConnection);
 		timer->deleteLater();
 	}
 
@@ -144,52 +146,8 @@ void ScriptExecutionControl::removeFile(const QString &file)
 	out.remove();
 }
 
-static inline int32_t getMedian(uint8_t &a, uint8_t &b, uint8_t &c, uint8_t &d)
-{
-	if (a > b)
-		std::swap(a, b);
-	if (c > d)
-		std::swap(c, d);
-	if (a > c)
-		std::swap(a, c);
-	if (b > d)
-		std::swap(b, d);
-	return (static_cast<int32_t>(b) + c) >> 1;
-}
-
 QVector<int32_t> ScriptExecutionControl::getPhoto()
 {
-	QLOG_INFO() << "Calling getStillImage()";
-	auto data = mBrick.getStillImage();
-	QVector<int32_t> result;
-	result.reserve(data.size() / 3); //Repack RGB88 from 3 x uint8_t into int32_t
-	constexpr auto IMAGE_WIDTH = 320;
-	constexpr auto IMAGE_HEIGHT = 240;
-	if (data.size() >= IMAGE_WIDTH * IMAGE_HEIGHT * 3) {
-		for(int row = 0; row < IMAGE_HEIGHT; row += 2) {
-			for(int col = 0; col < IMAGE_WIDTH; col+=2) {
-				auto row1 = &data[(row*IMAGE_WIDTH+col)*3];
-				auto row2 = row1 + IMAGE_WIDTH*3;
-				auto r1 = row1[0];
-				auto g1 = row1[1];
-				auto b1 = row1[2];
-				auto r2 = row1[3];
-				auto g2 = row1[4];
-				auto b2 = row1[5];
-				auto r3 = row2[0];
-				auto g3 = row2[1];
-				auto b3 = row2[2];
-				auto r4 = row2[3];
-				auto g4 = row2[4];
-				auto b4 = row2[5];
-
-				result.push_back((getMedian(r1, r2, r3, r4) << 16)
-					| (getMedian(g1, g2, g3, g4) << 8)
-					| getMedian(b1, b2, b3, b4));
-			}
-		}
-	}
-
-	QLOG_INFO() << "Constructed result of getStillImage()";
+	QVector<int32_t> result = trikControl::Utilities::rescalePhoto(mBrick.getStillImage());
 	return result;
 }
